Feat: 리액션 취소 예외처리

+++ BACKEND/JaraMe/src/main/java/siliconDream/jaraMe/service/ReactionServiceImpl.java
@@ -1,5 +1,6 @@
 package siliconDream.jaraMe.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -11,8 +12,11 @@ import siliconDream.jaraMe.repository.MissionPostRepository;
 import siliconDream.jaraMe.repository.ReactionRepository;
 import siliconDream.jaraMe.repository.UserRepository;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @Service
 public class ReactionServiceImpl implements ReactionService {
 
@@ -33,10 +37,23 @@ public class ReactionServiceImpl implements ReactionService {
     }
 
     public ResponseEntity<String> addReaction(MissionReactionDTO missionReactionDTO, Long userId) {
-        //TODO: 예외처리 : 존재하는 리액션 타입이 맞는지?
+        //가능한 reactionType
+        List<String> reactionTypes = new ArrayList<>();
+        reactionTypes.add("like");
+        reactionTypes.add("good");
+        reactionTypes.add("smile");
 
-        Optional<Reaction> reactionOptional=reactionRepository.findReactionByMissionPost_MissionPostIdAndUser_UserId(missionReactionDTO.getMissionPostId(), userId);
-        if (reactionOptional.isEmpty()){
+
+        Optional<Reaction> reactionOptional = reactionRepository.findReactionByMissionPost_MissionPostIdAndUser_UserId(missionReactionDTO.getMissionPostId(), userId);
+        if (reactionOptional.isEmpty()) {
+            //예외처리 : 리액션 타입 잘못 전달받았을 때 (ex- likeeee, gosssod, smiile )
+            if (!reactionTypes.contains(missionReactionDTO.getReactionType())) {
+
+                log.info("log test : test success");
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("리액션 타입이 잘못되었습니다.");
+
+            }
+
             Reaction reaction = new Reaction();
             reaction.setUser(userRepository.findByUserId(userId));
             reaction.setMissionPost(missionPostRepository.findByMissionPostId(missionReactionDTO.getMissionPostId()));
@@ -51,26 +68,39 @@ public class ReactionServiceImpl implements ReactionService {
                     missionReactionDTO.getReactionType());
             notificationService.createNotification(post.getUser().getUserId(), notificationMessage);
 
-            return ResponseEntity.status(HttpStatus.OK).body(String.format("%s를 누르셨습니다!",missionReactionDTO.getReactionType()));
-        }else if (reactionOptional.isPresent()){
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body( String.format("이미 %s를 누르셨습니다.",reactionOptional.get().getReactionType()));
+            return ResponseEntity.status(HttpStatus.OK).body(String.format("%s를 누르셨습니다!", missionReactionDTO.getReactionType()));
+        } else if (reactionOptional.isPresent()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(String.format("이미 %s를 누르셨습니다.", reactionOptional.get().getReactionType()));
         }
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("리액션을 누를 수 없습니다.");
 
     }
 
     public String deleteReaction(MissionReactionDTO missionReactionDTO, Long userId) {
+//가능한 reactionType
+        List<String> reactionTypes = new ArrayList<>();
+        reactionTypes.add("like");
+        reactionTypes.add("good");
+        reactionTypes.add("smile");
 
-        Optional<Reaction> reactionOptional=reactionRepository.findReactionByMissionPost_MissionPostIdAndUser_UserId(missionReactionDTO.getMissionPostId(), userId);
-        if (reactionOptional.isPresent()){
+        Optional<Reaction> reactionOptional = reactionRepository.findReactionByMissionPost_MissionPostIdAndUser_UserId(missionReactionDTO.getMissionPostId(), userId);
+        if (reactionOptional.isPresent()) {
 
-            reactionRepository.delete(reactionOptional.get());
-            return "리액션이 취소되었습니다.";
+            //예외처리 : 리액션 타입 잘못 전달받았을 때 (ex- likeeee, gosssod, smiile )
+            if (!reactionTypes.contains(missionReactionDTO.getReactionType())) {
 
-            //삭제가 됐는지 점검하는 부분
+                log.info("log test : test success");
+                return "리액션 타입이 잘못되었습니다.";
 
-        }   Optional<Reaction> deleteTestReactionOptional = reactionRepository.findByReactionId(reactionOptional.get().getReactionId());
-        if (deleteTestReactionOptional.isPresent()){return "리액션 취소에 실패했습니다.";            }
-    return "리액션 취소에 실패했습니다.";
+            } else if (!reactionOptional.get().getReactionType().equals(missionReactionDTO.getReactionType())) {
+                return "리액션 타입이 잘못되었습니다.";
+            } else {
+                reactionRepository.delete(reactionOptional.get());
+                return "리액션이 취소되었습니다.";
+
+            }
+        } else {
+            return "리액션 취소에 실패했습니다.";
+        }
     }
 }
